+++ source/android/uitestapp/src/androidTest/java/io.adaptivecards.uitestapp/UiTests.kt
@@ -91,8 +91,6 @@ class UiTests {
 
         TestHelpers.clickOnElementWithText("OK")
 
-        Assert.assertFalse(true);
-
         TestHelpers.goToInputsScreen()
         TestHelpers.assertInputValuePairExists("chosenAnimal", "Crimson Shining Parrot")
     }
